chore: use one function to perform UST in test scripts
